+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - On X11, fixed panic caused by dropping the window before running the event loop.
 - Introduce `WindowBuilderExt::with_app_id` to allow setting the application ID on Wayland.
-- On Windows, catch panics in event loop child thread and forward them to the parent thread.
+- On Windows, catch panics in event loop child thread and forward them to the parent thread. This prevents an invocation of undefined behavior due to unwinding into foreign code.
 
 # Version 0.18.0 (2018-11-07)